feat:rework folder store. prevent mv reordering
Turns out vue draggable works in ways I don't understand.

It does not in fact update the `position` property when it is set
as item-key, and even without the `position` property in files and
folders it still somehow knows how to order things. Very confusing.

Using a Map<string, VaunchFolder> for the folder store was messing
up how vuedraggable did ordering, so reworked the entire thing to
use a simple Array<VaunchFolder> instead. Slightly less efficient
when getting items, but makes things slightly easier to work with,
and allows us to just update the folder name when performing `mv`
as we don't need to worry about updating the key in a Map.

Thankfully everything was using calls from the folder store itself
so just updating the store methods to work with the array was all
that was really needed.

No changes to the serialised data was made, so no breaking changes
in that area, thankfully